refactor: BallAndStrike 클래스 분리
